requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
python-jose[cryptography]
python-multipart


